+++ astro.config.mjs
@@ -1,4 +1,9 @@
 import { defineConfig } from 'astro/config';
+import sitemap from "@astrojs/sitemap";
+import tailwind from "@astrojs/tailwind";
 
 // https://astro.build/config
-export default defineConfig({});
+export default defineConfig({
+  site: 'https://swinlead.com',
+  integrations: [sitemap(), tailwind()]
+});